Fix docker-compose
The __template__ service wasn't working, so I made it such that it works when no docker-compose.yaml.template is provided.

I then also took the opportunity to trim down what was in the other docker-compose.yaml.template files and updated the script to still work in those scenarios as well.